+++ starkbank/__init__.py
@@ -1,4 +1,4 @@
-version = "2.19.1"
+version = "2.20.0"
 
 user = None
 language = "en-US"